fix encoding problem in windows console
The file was save in an UTF8 encoding and (apparently) not converted back to a codepage suitable for the windows command line (codepage 850). This patch makes sure you get the right character independent of the file encoding. Obviously detecting the environment and serving up the right codepage would be nicer, but i recommend to leave this patch in for now until there is a better solution.